# Ques_backend/docs/api/api_specification.yaml
openapi: 3.0.3
info:
  title: Ques Backend API
  description: |
    Complete API specification for the Ques application backend.
    Includes AI-powered user discovery, chat system, swipe mechanics, payments, and more.
  version: 1.0.0
  contact:
    name: Ques Backend Team
    email: backend@ques.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.ques.app
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        bio:
          type: string
        phone_number:
          type: string
        wechat_id:
          type: string
        profile_picture_url:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
...
paths:
...
  /api/v1/users/profile:
    get:
      tags:
        - Users
      summary: Get user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'

    put:
      tags:
        - Users
      summary: Update user profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfile'
      responses:
        '200':
          description: Profile updated
  /api/v1/users/discover:
    get:
      tags:
        - Users
      summary: Discover users for swiping
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Users retrieved
  /api/v1/users/search:
    get:
      tags:
        - Users
      summary: Search users
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: skills
          in: query
          schema:
            type: array
            items:
              type: string
        - name: location
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Search results

  # Whisper endpoints
...